+++ src/cursor.ts
@@ -1,7 +1,4 @@
-const move = (code: string) => (count?: number): string =>
-	`\x1b[${count === undefined ? "" : count ? count : "?"}${code}`;
-
-const noop = (): string => "";
+import {csi, count as move, noop} from "./utils";
 
 const cursor = {
 	up       : move("A"),
@@ -11,35 +8,38 @@ const cursor = {
 	nextLine : move("E"),
 	prevLine : move("F"),
 
-	x: (x: number): string => `\x1b[${x}G`,
+	// Don't use `move`/`count` here, because it will output an empty string if `x` is `0`, but the
+	// terminal accepts `\x1b[0G` and will process it the same as `\x1b[1G`/`x1b[G`. Also, the
+	// parameter for `move`/`count` is optional, but the parameter for `x` is required.
+	x: (x: number) => csi(x, "G"),
 
 	// These are the old VT100 codes, but they're widely supported by old and new terminals. The
 	// newer VT220 codes are standardized by ECMA and ISO, and are also widely supported, but might
 	// not work in some older terminals. The VT220 codes are:
 	//
 	// ```
-	// save    : "\x1b[s",
-	// restore : "\x1b[u",
+	// save    : csi("s"),
+	// restore : csi("u"),
 	// ```
 	save    : "\x1b7",
 	restore : "\x1b8",
 
-	hide : "\x1b[?25l",
-	show : "\x1b[?25h",
+	hide : csi("?25l"),
+	show : csi("?25h"),
 
 	position: {
-		get: "\x1b[6n",
-		set: (x: number, y: number): string => `\x1b[${y};${x}H`,
+		get: csi("6n"),
+		set: (x: number, y: number): string => csi(`${y};${x}H`),
 	},
 
 	shape: {
-		steadyBlock       : "\x1b[2 q",
-		steadyBar         : "\x1b[6 q",
-		steadyUnderline   : "\x1b[4 q",
-		blinkingBlock     : "\x1b[1 q",
-		blinkingBar       : "\x1b[5 q",
-		blinkingUnderline : "\x1b[3 q",
-		default           : "\x1b[0 q",
+		steadyBlock       : csi("2 q"),
+		steadyBar         : csi("6 q"),
+		steadyUnderline   : csi("4 q"),
+		blinkingBlock     : csi("1 q"),
+		blinkingBar       : csi("5 q"),
+		blinkingUnderline : csi("3 q"),
+		default           : csi("0 q"),
 	},
 };