+++ src/scroll.ts
@@ -1,13 +1,15 @@
+import {count, noop} from "./utils";
+
 const scroll = {
-	up   : (lines?: number): string => `\x1b[${lines ?? 1}S`,
-	down : (lines?: number): string => `\x1b[${lines ?? 1}T`,
+	up   : count<[lines?: number]>("S"),
+	down : count<[lines?: number]>("T"),
 };
 
 export type Scroll = typeof scroll;
 
 const disabled: Scroll = {
-	up   : () => "",
-	down : () => "",
+	up   : noop,
+	down : noop,
 };
 
 export const makeScroll = (enabled = true): Scroll => enabled ? scroll : disabled;